requirements.txt
requests==2.18.2
lxml==4.1.1
Pillow==4.3.0